docs: Update the readme
Update the details on how to use the plugin
 - add info about auth
 - add more examples
