--- pkg/flag/aqua.go
@@ -1,10 +1,0 @@
-package flag
-
-import "github.com/spf13/cobra"
-
-func AddMcpFlags(cmd *cobra.Command) {
-	cmd.Flags().StringVarP(&transport, "transport", "t", "stdio", "Transport protocol to use (sse or stdio)")
-	cmd.Flags().IntVarP(&port, "port", "p", 23456, "Port to listen on")
-	cmd.Flags().StringVarP(&trivyBinary, "trivy-binary", "", "", "Path to the Trivy binary")
-	cmd.Flags().BoolVarP(&useAquaPlatform, "use-aqua-platform", "a", false, "Use Aqua platform")
-}
