+++ pkg/flag/flag.go
@@ -15,8 +15,14 @@ var (
 	aquaKey    string
 	aquaSecret string
 	aquaRegion string
-	clear      bool
 )
+
+func AddMcpFlags(cmd *cobra.Command) {
+	cmd.Flags().StringVarP(&transport, "transport", "t", "stdio", "Transport protocol to use (sse or stdio)")
+	cmd.Flags().IntVarP(&port, "port", "p", 23456, "Port to listen on")
+	cmd.Flags().StringVarP(&trivyBinary, "trivy-binary", "", "", "Path to the Trivy binary")
+	cmd.Flags().BoolVarP(&useAquaPlatform, "use-aqua-platform", "a", false, "Use Aqua platform")
+}
 
 func AddBaseFlags(cmd *cobra.Command) {
 	cmd.PersistentFlags().BoolVarP(&showVersion, "version", "v", false, "Show version")
@@ -27,7 +33,6 @@ func AddLoginFlags(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&aquaKey, "aqua-key", "", "Aqua key")
 	cmd.Flags().StringVar(&aquaSecret, "aqua-secret", "", "Aqua secret")
 	cmd.Flags().StringVar(&aquaRegion, "aqua-region", "", "Aqua region (US, EU, Singapore, Sydney, Dev)")
-	cmd.Flags().BoolVar(&clear, "clear", false, "Clear the saved credentials")
 }
 
 func ToOptions() Options {
@@ -48,6 +53,5 @@ func ToLoginOptions() LoginOptions {
 		AquaKey:    aquaKey,
 		AquaSecret: aquaSecret,
 		AquaRegion: aquaRegion,
-		Clear:      clear,
 	}
 }
